tubers/requirement.txt
asgiref==3.3.4
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.2.4
django-allauth==0.44.0
django-ckeditor==6.1.0
django-heroku==0.3.1
django-js-asset==1.2.2
django-on-heroku==1.1.2
djangocms-admin-style==2.0.2
gunicorn==20.1.0
idna==2.10
mysqlclient==2.0.3
oauthlib==3.1.1
Pillow==8.2.0
psycopg2==2.8.6
psycopg2-binary==2.8.6
pycparser==2.20
PyJWT==2.1.0
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
sqlparse==0.4.1
urllib3==1.26.5
whitenoise==5.2.0
